+++ CHANGELOG.md
@@ -1,11 +1,204 @@
-## [6.0.2](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.1...v6.0.2) (2025-01-15)
+# [7.0.0-next.10](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.9...v7.0.0-next.10) (2025-01-20)
 
 
 ### Bug Fixes
 
 * **react:** [sort-jsx-props] use regex pattern for custom groups ([8625bbd](https://github.com/Boehringer-Ingelheim/eslint-config/commit/8625bbde24df50944fd661b70d50e2ae9cfe4217))
 
-## [6.0.1](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.0...v6.0.1) (2024-12-18)
+# [7.0.0-next.9](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.8...v7.0.0-next.9) (2025-01-17)
+
+
+### Features
+
+* **naming-convention:** adds a new shared configuration for enforcing naming conventions ([8f2fdf2](https://github.com/Boehringer-Ingelheim/eslint-config/commit/8f2fdf2afee4aed0a26ab4f3a98d22627daefe5b))
+
+# [7.0.0-naming-convetion-configuration.2](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-naming-convetion-configuration.1...v7.0.0-naming-convetion-configuration.2) (2025-01-17)
+
+
+### Bug Fixes
+
+* **base:** turn off import/no-unused-modules for now ([a8f15dc](https://github.com/Boehringer-Ingelheim/eslint-config/commit/a8f15dc3a35fc6e6d774d8e766985772009862f0))
+* publishes type declarations ([a410e8e](https://github.com/Boehringer-Ingelheim/eslint-config/commit/a410e8eed04e5bde28eb6aa14be19858bb685cc9))
+
+# [7.0.0-naming-convetion-configuration.1](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.1...v7.0.0-naming-convetion-configuration.1) (2025-01-17)
+
+
+### Features
+
+* **naming-convention:** adds a new shared configuration for enforcing naming conventions ([8f2fdf2](https://github.com/Boehringer-Ingelheim/eslint-config/commit/8f2fdf2afee4aed0a26ab4f3a98d22627daefe5b))
+
+# [7.0.0-next.8](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.7...v7.0.0-next.8) (2025-01-17)
+
+
+### Bug Fixes
+
+* publishes type declarations ([a410e8e](https://github.com/Boehringer-Ingelheim/eslint-config/commit/a410e8eed04e5bde28eb6aa14be19858bb685cc9))
+
+# [7.0.0-next.7](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.6...v7.0.0-next.7) (2025-01-17)
+
+
+### Bug Fixes
+
+* **base:** turn off import/no-unused-modules for now ([a8f15dc](https://github.com/Boehringer-Ingelheim/eslint-config/commit/a8f15dc3a35fc6e6d774d8e766985772009862f0))
+
+# [7.0.0-next.6](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.5...v7.0.0-next.6) (2025-01-15)
+
+
+### Bug Fixes
+
+* **react:** [sort-jsx-props] use regex pattern for custom groups ([720df19](https://github.com/Boehringer-Ingelheim/eslint-config/commit/720df196a0b6f1e0c480f95b7e6c6fa47e82bac0))
+
+
+### Features
+
+* **deps:** update dependencies ([564b3ba](https://github.com/Boehringer-Ingelheim/eslint-config/commit/564b3ba9aec9ad1dbc80477438abfdfd82f69c6a))
+
+# [6.0.2](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.1...v6.0.2) (2025-01-15)
+
+
+### Bug Fixes
+
+* **react:** [sort-jsx-props] use regex pattern for custom groups ([8625bbd](https://github.com/Boehringer-Ingelheim/eslint-config/commit/8625bbde24df50944fd661b70d50e2ae9cfe4217))
+
+# [7.0.0-next.5](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.4...v7.0.0-next.5) (2025-01-09)
+
+
+### Bug Fixes
+
+* **shared-configurations:** [nextjs] ignore .next folder per default ([2f51caa](https://github.com/Boehringer-Ingelheim/eslint-config/commit/2f51caaf8b7ead38db9d753586f930ae91a8e3e5))
+* **shared-configurations:** [react] remove redefined jsx-a11y plugin ([76e7fc4](https://github.com/Boehringer-Ingelheim/eslint-config/commit/76e7fc4f06caf252cc5bbcdbeb8fb4848c748e93))
+* **shared-configurations:** make next.js configuration importable ([c6ca951](https://github.com/Boehringer-Ingelheim/eslint-config/commit/c6ca95150866af2c045a73b706ac3f034f34d7b6))
+
+
+### Features
+
+* **shared-configurations:** [nextjs] enhance compatibility with react-refresh ([ba20ebd](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ba20ebdced704db40602aa25d270496c06d18d7d))
+* **shared-configurations:** add next.js configuration ([ca94532](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ca9453226d898196c576073adc7af47869b551a4))
+
+# [7.0.0-shared-configuration-nextjs.4](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-shared-configuration-nextjs.3...v7.0.0-shared-configuration-nextjs.4) (2025-01-08)
+
+
+### Bug Fixes
+
+* **shared-configurations:** [nextjs] ignore .next folder per default ([2f51caa](https://github.com/Boehringer-Ingelheim/eslint-config/commit/2f51caaf8b7ead38db9d753586f930ae91a8e3e5))
+
+# [7.0.0-shared-configuration-nextjs.3](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-shared-configuration-nextjs.2...v7.0.0-shared-configuration-nextjs.3) (2025-01-08)
+
+
+### Bug Fixes
+
+* **shared-configurations:** [react] remove redefined jsx-a11y plugin ([76e7fc4](https://github.com/Boehringer-Ingelheim/eslint-config/commit/76e7fc4f06caf252cc5bbcdbeb8fb4848c748e93))
+
+
+### Features
+
+* **shared-configurations:** [nextjs] enhance compatibility with react-refresh ([ba20ebd](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ba20ebdced704db40602aa25d270496c06d18d7d))
+
+# [7.0.0-shared-configuration-nextjs.2](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-shared-configuration-nextjs.1...v7.0.0-shared-configuration-nextjs.2) (2025-01-08)
+
+
+### Bug Fixes
+
+* **shared-configurations:** make next.js configuration importable ([c6ca951](https://github.com/Boehringer-Ingelheim/eslint-config/commit/c6ca95150866af2c045a73b706ac3f034f34d7b6))
+
+# [7.0.0-shared-configuration-nextjs.1](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.1...v7.0.0-shared-configuration-nextjs.1) (2025-01-08)
+
+
+### Bug Fixes
+
+* **base:** switch some perfectionist references to flatconfig ([a6697cc](https://github.com/Boehringer-Ingelheim/eslint-config/commit/a6697cc2c6c11ff717732abe741d1300c6d3a608))
+* **scripts:** [repair] use exact rimraf version to not use a local and possible incompatible version ([059835d](https://github.com/Boehringer-Ingelheim/eslint-config/commit/059835d26311ea30a6dd5e837eb1a040dcad2b69))
+
+
+### Features
+
+* add basic typing to the index.js ([ce3fbb0](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ce3fbb063150a162085df2f6e110460670612bb1))
+* **base:** [typescript-eslint] update parserOptions to use projectService instead of project ([02ac930](https://github.com/Boehringer-Ingelheim/eslint-config/commit/02ac930eca6f1d9da4f2c724c1febebd9f3c2524))
+* **deps:** update dependencies ([dfed600](https://github.com/Boehringer-Ingelheim/eslint-config/commit/dfed60091dfc8be82b1290841fbb864eda982762))
+* migrates to flat-config ([b22a872](https://github.com/Boehringer-Ingelheim/eslint-config/commit/b22a872a1589516d571c8907e6ec30bb2b529f94))
+* **shared-configurations:** add next.js configuration ([ca94532](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ca9453226d898196c576073adc7af47869b551a4))
+
+
+### BREAKING CHANGES
+
+* migrate to flat config
+* drop support for legacy config
+
+# [7.0.0-next.4](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.3...v7.0.0-next.4) (2025-01-08)
+
+
+### Bug Fixes
+
+* **scripts:** [repair] use exact rimraf version to not use a local and possible incompatible version ([059835d](https://github.com/Boehringer-Ingelheim/eslint-config/commit/059835d26311ea30a6dd5e837eb1a040dcad2b69))
+
+
+### Features
+
+* **deps:** update dependencies ([dfed600](https://github.com/Boehringer-Ingelheim/eslint-config/commit/dfed60091dfc8be82b1290841fbb864eda982762))
+
+# [7.0.0-next.3](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.2...v7.0.0-next.3) (2025-01-08)
+
+
+### Features
+
+* **base:** [typescript-eslint] update parserOptions to use projectService instead of project ([02ac930](https://github.com/Boehringer-Ingelheim/eslint-config/commit/02ac930eca6f1d9da4f2c724c1febebd9f3c2524))
+
+# [7.0.0-typescript-eslint-parser-options.1](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.1...v7.0.0-typescript-eslint-parser-options.1) (2025-01-07)
+
+
+### Bug Fixes
+
+* **base:** switch some perfectionist references to flatconfig ([a6697cc](https://github.com/Boehringer-Ingelheim/eslint-config/commit/a6697cc2c6c11ff717732abe741d1300c6d3a608))
+
+
+### Features
+
+* add basic typing to the index.js ([ce3fbb0](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ce3fbb063150a162085df2f6e110460670612bb1))
+* **base:** [typescript-eslint] update parserOptions to use projectService instead of project ([02ac930](https://github.com/Boehringer-Ingelheim/eslint-config/commit/02ac930eca6f1d9da4f2c724c1febebd9f3c2524))
+* migrates to flat-config ([b22a872](https://github.com/Boehringer-Ingelheim/eslint-config/commit/b22a872a1589516d571c8907e6ec30bb2b529f94))
+
+
+### BREAKING CHANGES
+
+* migrate to flat config
+* drop support for legacy config
+
+# [7.0.0-next.2](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v7.0.0-next.1...v7.0.0-next.2) (2024-12-20)
+
+
+### Bug Fixes
+
+* **base:** switch some perfectionist references to flatconfig ([a6697cc](https://github.com/Boehringer-Ingelheim/eslint-config/commit/a6697cc2c6c11ff717732abe741d1300c6d3a608))
+
+# [7.0.0-next.1](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.1...v7.0.0-next.1) (2024-12-19)
+
+
+### Features
+
+* add basic typing to the index.js ([ce3fbb0](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ce3fbb063150a162085df2f6e110460670612bb1))
+* migrates to flat-config ([b22a872](https://github.com/Boehringer-Ingelheim/eslint-config/commit/b22a872a1589516d571c8907e6ec30bb2b529f94))
+
+
+### BREAKING CHANGES
+
+* migrate to flat config
+* drop support for legacy config
+
+# [7.0.0-migrate-flat-config.1](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.1...v7.0.0-migrate-flat-config.1) (2024-12-18)
+
+
+### Features
+
+* add basic typing to the index.js ([ce3fbb0](https://github.com/Boehringer-Ingelheim/eslint-config/commit/ce3fbb063150a162085df2f6e110460670612bb1))
+* migrates to flat-config ([b22a872](https://github.com/Boehringer-Ingelheim/eslint-config/commit/b22a872a1589516d571c8907e6ec30bb2b529f94))
+
+
+### BREAKING CHANGES
+
+* migrate to flat config
+* drop support for legacy config
+
+# [6.0.1](https://github.com/Boehringer-Ingelheim/eslint-config/compare/v6.0.0...v6.0.1) (2024-12-18)
 
 
 ### Bug Fixes